refactor: move defineCollections file
