+++ imports/plugins/core/files/server/fileCollections.js
@@ -12,7 +12,7 @@ import {
 } from "@reactioncommerce/file-collections";
 import { AbsoluteUrlMixin } from "/imports/plugins/core/core/server/Reaction/absoluteUrl";
 import { MediaRecords } from "/lib/collections";
-import setUpFileCollections from "/imports/services/files/setUpFileCollections";
+import setUpFileCollections from "/imports/node-app/services/files/setUpFileCollections";
 
 /**
  * Functions and objects related to file upload, download, and storage
