+++ tests/TestApp.js
@@ -4,11 +4,11 @@ import findFreePort from "find-free-port";
 import MongoDBMemoryServer from "mongodb-memory-server";
 import Random from "@reactioncommerce/random";
 import appEvents from "../imports/plugins/core/core/server/appEvents";
-import createApolloServer from "../imports/plugins/core/graphql/server/no-meteor/createApolloServer";
-import defineCollections from "../imports/collections/defineCollections";
+import createApolloServer from "../imports/node-app/core/createApolloServer";
+import defineCollections from "../imports/node-app/core/util/defineCollections";
 import Factory from "../imports/test-utils/helpers/factory";
 import hashLoginToken from "../imports/node-app/core/util/hashLoginToken";
-import setUpFileCollections from "../imports/services/files/setUpFileCollections";
+import setUpFileCollections from "../imports/node-app/services/files/setUpFileCollections";
 
 class TestApp {
   constructor() {
